REFACTORIZACION de codigos vacios y llenos
