+++ src/JugadorA.js
@@ -1,0 +1,10 @@
+function codigoSecreto(codigo) {
+
+    if(codigo=="") {
+        return "Codigo Vacio";    
+    }
+    return codigo;
+    ;
+}
+
+export default codigoSecreto;
