+++ src/JugadorB.js
@@ -1,0 +1,6 @@
+function recibirCodigoB(codigo) {
+
+    return "Ingresaste codigo vacio";        
+}
+
+export default recibirCodigoB;
